requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-WTF==1.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
WTForms==3.0.1
pymysql==1.0.3
requests==2.32.3
Jinja2==3.1.2
Werkzeug==3.1.3
